feat(bill): implement search functionality and update UI
Implemented search functionality for bills and updated the UI.

- Added search functionality to `BillsViewModel` using `savedStateHandle` to persist the search query and widget state.
- Updated `BillsScreen` to use `SearchTopAppBar` and `SearchToolbar` for the search functionality.
- Updated `BillsList` to filter bills based on the search query.
- Added `BillsTopAppBar` composable that manages search and default app bar.
- Updated `BillItem` to display due amounts.
- Updated `BillsScreen` to use the new `BillsTopAppBar` composable.
- Updated `BillsList` to use a draggable scrollbar
